переносит функцию генерации объекта и константы в data и вспомогательные функции в util
